+++ R/online_ODRF.R
@@ -5,6 +5,7 @@
 #' @param X An new n by d numeric matrix (preferable) or data frame  used to update the object of class \code{ODRF}.
 #' @param y A new response vector of length n used to update the object of class \code{ODRF}.
 #' @param weights A vector of non-negative observational weights; fractional weights are allowed (default NULL).
+#' @param MaxDepth The maximum depth of the tree (default \code{Inf}).
 #' @param ... Optional parameters to be passed to the low level function.
 #'
 #' @return The same result as \code{ODRF}.
@@ -52,35 +53,35 @@
 #' @aliases online.ODRF
 #' @method online ODRF
 #' @export
-online.ODRF <- function(obj, X, y, weights = NULL, ...) {
-  ppForest <- obj
-  rm(obj)
-  if (length(ppForest[["ppTrees"]][[1]][["structure"]][["nodeDepth"]]) == 1) {
+online.ODRF <- function(obj, X, y, weights = NULL, MaxDepth = Inf, ...) {
+  if (length(obj[["structure"]][[1]][["structure"]][["nodeDepth"]]) == 1) {
     stop("No tree structure to use 'online'!")
   }
+
+  MaxDepth0 <- MaxDepth
   weights0 <- weights
-  ppTrees <- ppForest$ppTrees
-  Levels <- ppForest$Levels
-  split <- ppForest$split
-  NodeRotateFun <- ppForest$NodeRotateFun
-  Call <- ppForest$call
-  Terms <- ppForest$terms
-  paramList <- ppForest$paramList
+  structure <- obj$structure
+  Levels <- obj$Levels
+  split <- obj$split
+  NodeRotateFun <- obj$NodeRotateFun
+  Call <- obj$call
+  Terms <- obj$terms
+  paramList <- obj$paramList
 
   subset <- weights <- na.action <- n <- p <- varName <- Xscale <- minCol <- maxminCol <- Xcat <- catLabel <- NULL
   lambda <- FunDir <- MaxDepth <- MinLeaf <- numNode <- TreeRandRotate <- NULL
-  ntrees <- numOOB <- storeOOB <- replacement <- stratify <- parallel <- numCores <- NULL
+  ntrees <-  ratOOB <- storeOOB <- replacement <- stratify <- parallel <- numCores <- NULL
 
-  ppForest <- ppForest[c(9, 10, 11)]
-  ppForestVar <- c(names(ppForest$data), names(ppForest$tree), names(ppForest$forest))
-  ppForest <- do.call("c", ppForest)
+  obj <- obj[(length(obj)-(3:1))]
+  ppForestVar <- c(names(obj$data), names(obj$tree), names(obj$forest))
+  obj <- do.call("c", obj)
 
   # env <- new.env()
   for (v in seq_along(ppForestVar)) {
-    assign(ppForestVar[v], ppForest[[v]]) # ,envir = env)
+    assign(ppForestVar[v], obj[[v]]) # ,envir = env)
   }
-  rm(ppForest)
-  # utils::globalVariables(ppForestVar)
+  rm(obj)
+  # utils::globalVariables(objVar)
 
   # vars=all.vars(Terms)
   if (sum(Xcat) > 0 && is.null(catLabel)) {
@@ -111,25 +112,17 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
   # if(!is.null(weights))
   #  X <- X * matrix(weights0,length(y),ncol(X))
   weights <- weights0
-
+  MaxDepth <- MaxDepth0
 
   ppForest <- list(
-    call = Call, terms = Terms, split = split, Levels = NULL,
-    NodeRotateFun = NodeRotateFun, paramList = paramList, oobErr = NULL, oobConfusionMat = NULL
+    call = Call, terms = Terms, split = split, Levels = NULL, NodeRotateFun = NodeRotateFun,
+    predicted=NULL, paramList = paramList, oobErr = NULL, oobConfusionMat = NULL
   )
+
   if (split != "mse") {
-    # adjust y to go from 1 to numClass if needed
-    if (is.factor(y)) {
-      ppForest$Levels <- levels(y)
-      y <- as.integer(y)
-    } else if (is.numeric(y)) {
-      ppForest$Levels <- sort(unique(y))
-      y <- as.integer(as.factor(y))
-    } else {
-      ppForest$Levels <- levels(as.factor(y))
-      y <- as.integer(as.factor(y))
-      # stop("Incompatible X type. y must be of type factor or numeric.")
-    }
+    y <- as.factor(y)
+    ppForest$Levels <- levels(y)
+    y <- as.integer(y)
 
     numClass <- length(ppForest$Levels)
     classCt <- cumsum(table(y))
@@ -187,19 +180,24 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
 
   ppForest$data <- list(
     subset = subset, weights = weights, na.action = na.action, n = n, p = p, varName = varName,
-    Xscale = Xscale, minCol = minCol, maxminCol = maxminCol, Xcat = Xcat, catLabel = catLabel
+    Xscale = Xscale, minCol = minCol, maxminCol = maxminCol, Xcat = Xcat, catLabel = catLabel, TreeRandRotate = TreeRandRotate
   )
-  ppForest$tree <- list(lambda = lambda, FunDir = FunDir, MaxDepth = MaxDepth, MinLeaf = MinLeaf, numNode = numNode, TreeRandRotate = TreeRandRotate)
+  ppForest$tree <- list(lambda = lambda, FunDir = FunDir, MaxDepth = MaxDepth, MinLeaf = MinLeaf, numNode = numNode)
   ppForest$forest <- list(
-    ntrees = ntrees, numOOB = numOOB, storeOOB = storeOOB, replacement = replacement, stratify = stratify,
+    ntrees = ntrees,  ratOOB =  ratOOB, storeOOB = storeOOB, replacement = replacement, stratify = stratify,
     parallel = parallel, numCores = numCores#, seed = seed
   )
 
 
   PPtree <- function(itree, ...) {
-    ppTree <- ppTrees[[itree]] # [1:7]
-    if ((numOOB > 0) && storeOOB) {
-      ppTree <- ppTree[-(length(ppTree) - c(2, 1, 0))]
+    ppTree=ppForest[seq(7)]
+    ppTree$data <- c(ppForest$data,structure[[itree]][c(1,2)])
+    ppTree$data$Xcat <- 0L
+    ppTree$data$Xscale <- "No"
+    ppTree$tree <- ppForest$tree
+    ppTree$structure <- structure[[itree]][-c(1,2)]
+    if (( ratOOB > 0) && storeOOB) {
+      ppTree$structure <- ppTree$structure[-(length(ppTree$structure) - c(2, 1, 0))]
     }
     class(ppTree) <- "ODT"
     #set.seed(seed + itree)
@@ -225,13 +223,14 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
         go <- all(TDindx0 %in% TDindx)
       }
     } else {
-      TDindx <- sample.int(TDindx0, n - numOOB, replace = FALSE)
+      TDindx <- sample.int(TDindx0, ceiling(n * (1 - ratOOB)), replace = FALSE)
     }
 
-    weights1 <- weights[TDindx]
-    ppForestT <- online(ppTree, X[TDindx, ], y[TDindx], weights1)
+    ppForestT <- online(ppTree, X[TDindx, ], y[TDindx], weights[TDindx])
 
-    if ((numOOB > 0) && storeOOB) {
+    TreeRotate=list(rotdims=ppForestT[["data"]][["rotdims"]],rotmat=ppForestT[["data"]][["rotmat"]])
+
+    if (( ratOOB > 0) && storeOOB) {
       oobErr <- 1
       NTD <- setdiff(TDindx0, TDindx)
       pred <- predict(ppForestT, X[NTD, ])
@@ -242,10 +241,12 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
         oobErr <- mean((pred - y[NTD])^2)
       }
 
-      ppForestT <- c(ppForestT, list(oobErr = oobErr, oobIndex = NTD, oobPred = pred))
+      ppForestT <- c(ppForestT$structure, list(oobErr = oobErr, oobIndex = NTD, oobPred = pred))
+    }else{
+      ppForestT <- ppForestT$structure
     }
 
-    return(ppForestT)
+    return(c(TreeRotate,ppForestT))
   }
 
 
@@ -268,8 +269,8 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
     # set.seed(seed)
     icore <- NULL
     ppForestT <- foreach::foreach(
-      icore = seq_along(chunks), .combine = list, .multicombine = TRUE, .export = c("ODT.compute"),
-      .packages = "ODRF", .noexport = "ppForest"
+      icore = seq_along(chunks), .combine = list, .multicombine = TRUE, .export = c("ODT_compute"),
+      .packages = "ODRF"
     ) %dopar% {
       lapply(chunks[[icore]], PPtree)
     }
@@ -277,22 +278,22 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
     parallel::stopCluster(cl)
 
     # do.call(rbind.fill,list1)
-    ppForest$ppTrees <- do.call("c", ppForestT)
-    # ppForest$ppTrees=NULL
+    ppForest$structure <- do.call("c", ppForestT)
+    # ppForest$structure=NULL
     # for (i in 1:numCores) {
-    #  ppForest$ppTrees=c(ppForest$ppTrees,ppForestT[[i]])
+    #  ppForest$structure=c(ppForest$structure,ppForestT[[i]])
     # }
   } else {
     # Use just one core.
-    ppForest$ppTrees <- lapply(1:ntrees, PPtree)
+    ppForest$structure <- lapply(1:ntrees, PPtree)
   }
 
 
   ####################################
-  if ((numOOB > 0) && storeOOB) {
+  if (( ratOOB > 0) && storeOOB) {
     oobVotes <- matrix(NA, n, ntrees)
     for (t in seq_len(ntrees)) {
-      oobVotes[ppForest$ppTrees[[t]]$oobIndex, t] <- ppForest$ppTrees[[t]]$oobPred
+      oobVotes[ppForest$structure[[t]]$oobIndex, t] <- ppForest$structure[[t]]$oobPred
     }
     idx <- which(rowSums(is.na(oobVotes)) < ntrees)
     oobVotes <- oobVotes[idx, , drop = FALSE]
@@ -326,6 +327,8 @@ online.ODRF <- function(obj, X, y, weights = NULL, ...) {
       oobPred <- rowMeans(oobVotes, na.rm = TRUE)
       ppForest$oobErr <- mean((oobPred - yy)^2) / mean((yy - mean(y))^2)
     }
+
+    ppForest$predicted <- oobPred
   }
 
   # class(ppTree) <- append(class(ppTree),"ODRF")