+++ tests/testthat/test-ODRF.R
@@ -4,9 +4,9 @@ dat <- data.matrix(iris)
 forest <- ODRF(Species ~ ., data = iris, split = "entropy", parallel = FALSE)
 
 ## Basic tests (for all random forests equal)
-test_that("classification iris forest is of class ODRF with 12 elements", {
+test_that("classification iris forest is of class ODRF with 13 elements", {
   expect_s3_class(forest, "ODRF")
-  expect_length(forest, 12)
+  expect_length(forest, 13)
 })
 
 test_that("Error if data is class matrix, classification", {