+++ NEWS.md
@@ -1,7 +1,15 @@
+# ODRF 0.0.4
+
+* Fixed function VarImp(), adding the method of measuring the importance of variables with node purity, and now VarImp() can be used for both class ODT and ODRF. 
+* When the argument “Xcat ! = 0”, i.e., the category variable in predictor X is transformed to one-of-K encode. however for the argument “NodeRotateFun=‘RotMatRF’ (‘RotMatRand’)“ run error, we have now fixed it. 
+* Added predicted values of training data for class ODT and ODRF.
+* Fixed issue related to function predict.ODRF() when argument "weight.tree = TRUE".
+* Optimized some other known issues.
+
 # ODRF 0.0.3
 
 * The function predicate.ODT() runs error when ODT is not split (depth=1), and we have fixed this bug.
-* We have fixed the function predict.ODRF with arguments numOOB and weight.tree related issues.
+* We have fixed the function predict.ODRF() with arguments numOOB and weight.tree related issues.
 * We have fixed the functions plot.ODT(), VarImp() and plot.VarImp().
 * We have fixed the argument 'lambda' of the functions ODT() and ODRF().